Implemented support for marker entries in undo changesets!
When undo changesets are transferred from buffer-undo-list to
buffer-undo-tree, any markers are moved into a weak hash table, and replaced
in the marker changeset element by a unique id that serves as the hash table
key. When the changeset is copied for passing to primitive-undo, the reverse
process replaces the id's by the corresponding markers from the hash
table. However, if a marker has been garbage collected in the meantime, it
will have been removed from the weak hash table, attempting to retrieve it
will return null, and the corresponding marker changeset element will not be
included in the copy of the changeset.