Added additional check to more reliably identify marker undo elements.
Recent bzr Emacs builds seem to add a new (and currently undocumented) buffer
modification undo element format: (t . -1) which appears to be used for the
initial modification entry for new files. This was getting incorrectly
identified as a marker undo entry, and deleted by undo-list-clean-GCd-elts,
preventing the buffer modification state from being restored correctly on
undoing.

The additional check in undo-list-GCd-marker-elt-p, on the symbol name in the
car of the undo element, fixes this.